// goPillion/node-service/trip-service/src/types/event.types.ts
export enum EventTypes {
    RIDE_CREATED = "ride_created",
    RIDE_REQUESTED = "ride_requested",
    RIDE_ACCEPTED = "ride_accepted",
    RIDE_CANCELLED = "ride_cancelled",
    RIDE_COMPLETED = "ride_completed",
    DRIVER_ARRIVED = "driver_arrived",
    OTP = "otp",
    CHAT_MESSAGE = "chat_message",
    SYSTEM_ALERT = "system_alert"
}